refactor(swagger): move FileDTO schema to its own file
